Working

+++ mqtt-sub-ctl/src/events.h
@@ -3,6 +3,7 @@
 
 #include "mail_helper.h"
 #include "mqtt_helper.h"
+#include <stdlib.h>
 
 #define EVENT_FIELD_LENGTH_MAX 256
 #define EVENT_CAP 16
@@ -15,7 +16,7 @@ struct event {
   char value[64];
   char sender_email[MAIL_FIELD_LENGTH_MAX];
   char recipient_emails[MAIL_MAX_RECIPIENTS][MAIL_FIELD_LENGTH_MAX];
-  int recipient_count;
+  size_t recipient_count;
 };
 
 #endif // EVENTS_H
